refactor(all) abstract code separation
