+++ module/category/src/Domain/ValueObject/CategoryCode.php
@@ -15,14 +15,6 @@ class CategoryCode extends AbstractCode
 {
     private const PATTERN = '/^[a-zA-Z0-9-_]+$\b/i';
 
-    public function __construct(string $value)
-    {
-        if (!self::isValid($value)) {
-            throw new \InvalidArgumentException(sprintf('Invalid category code value "%s"', $value));
-        }
-        parent::__construct($value);
-    }
-
     public static function isValid(string $value): bool
     {
         if (preg_match(self::PATTERN, $value, $matches) === 0) {
