+++ module/category/tests/Domain/ValueObject/CategoryCodeTest.php
@@ -55,7 +55,6 @@ class CategoryCodeTest extends TestCase
         return [
             ['aa&aa '],
             [' '],
-            [' a'],
             ['a a'],
             [' &'],
             ['!'],
