--- module/shared-kernel/src/Domain/ValueObject/Code.php
@@ -1,15 +1,0 @@
-<?php
-/*
- * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
- * See LICENSE.txt for license details.
- */
-
-declare(strict_types=1);
-
-namespace Ergonode\SharedKernel\Domain\ValueObject;
-
-use Ergonode\SharedKernel\Domain\AbstractCode;
-
-class Code extends AbstractCode
-{
-}
